requirements.txt
discord.py 
requests
beautifulsoup4
pillow
table2ascii